Move workos model
