+++ workos/resources/webhooks.py
@@ -4,7 +4,7 @@ from typing_extensions import Annotated
 from workos.types.directory_sync import DirectoryGroup
 from workos.types.events import EventPayload
 from workos.resources.user_management import OrganizationMembership, User
-from workos.resources.workos_model import WorkOSModel
+from workos.types.workos_model import WorkOSModel
 from workos.types.directory_sync.directory_user import DirectoryUser
 from workos.types.events.authentication_payload import (
     AuthenticationEmailVerificationSucceededPayload,
