+++ workos/types/directory_sync/directory.py
@@ -1,5 +1,5 @@
 from typing import Literal, Optional
-from workos.resources.workos_model import WorkOSModel
+from workos.types.workos_model import WorkOSModel
 from workos.types.directory_sync.directory_state import DirectoryState
 from workos.types.directory_sync.directory_type import DirectoryType
 from workos.typing.literals import LiteralOrUntyped
