+++ workos/types/events/authentication_payload.py
@@ -1,5 +1,5 @@
 from typing import Literal, Optional
-from workos.resources.workos_model import WorkOSModel
+from workos.types.workos_model import WorkOSModel
 
 
 class AuthenticationResultCommon(WorkOSModel):
