+++ workos/types/events/directory_payload_with_legacy_fields.py
@@ -1,5 +1,5 @@
 from typing import Literal, Sequence
-from workos.resources.workos_model import WorkOSModel
+from workos.types.workos_model import WorkOSModel
 from workos.types.events.directory_payload import DirectoryPayload
