+++ workos/types/events/event_model.py
@@ -1,6 +1,6 @@
 from typing import Generic, Literal, TypeVar
 from workos.resources.user_management import OrganizationMembership, User
-from workos.resources.workos_model import WorkOSModel
+from workos.types.workos_model import WorkOSModel
 from workos.types.directory_sync.directory_group import DirectoryGroup
 from workos.types.directory_sync.directory_user import DirectoryUser
 from workos.types.events.authentication_payload import (
