+++ workos/types/events/organization_domain_verification_failed_payload.py
@@ -1,5 +1,5 @@
 from typing import Literal
-from workos.resources.workos_model import WorkOSModel
+from workos.types.workos_model import WorkOSModel
 from workos.types.organizations.organization_domain import OrganizationDomain
 from workos.typing.literals import LiteralOrUntyped
