+++ workos/types/mfa/authentication_challenge_verification_response.py
@@ -1,4 +1,4 @@
-from workos.resources.workos_model import WorkOSModel
+from workos.types.workos_model import WorkOSModel
 from workos.types.mfa.authentication_challenge import AuthenticationChallenge
