+++ workos/types/mfa/authentication_factor_totp_and_challenge_response.py
@@ -1,4 +1,4 @@
-from workos.resources.workos_model import WorkOSModel
+from workos.types.workos_model import WorkOSModel
 from workos.types.mfa.authentication_challenge import AuthenticationChallenge
 from workos.types.mfa.authentication_factor import AuthenticationFactorTotpExtended
