+++ workos/types/organizations/organization_common.py
@@ -1,5 +1,5 @@
 from typing import Literal, Sequence
-from workos.resources.workos_model import WorkOSModel
+from workos.types.workos_model import WorkOSModel
 from workos.types.organizations.organization_domain import OrganizationDomain
